backend/requirements.txt
asgiref==3.2.10
django==2.2.16
django-filter==2.4.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
djoser==2.1.0
drf-extra-fields==3.4.0
flake8==5.0.4
gunicorn==20.0.4
Pillow==9.0.1
PyJWT==2.1.0
psycopg2-binary==2.8.6
python-dotenv==0.21.0
pytz==2022.6
requests==2.26.0